# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth-server",
        "title": "Auth-Server API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/user/login": {
            "post": {
                "description": "用户登录",
                "tags": [
                    "User"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "Json请求体",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tools.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户注册",
                "tags": [
                    "User"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "Json请求体",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tools.ResponseStruct"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.UserLoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "handlers.UserRegisterRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "tools.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据结构体"
                },
                "msg": {
                    "description": "相应说明",
                    "type": "string"
                }
            }
        }
    }
}